dataproduct_builder/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.jupytext]
formats = "ipynb,py:percent"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["F401"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git
    | \.venv
    | _build
    | build
    | dist
  )/
  | paths.py           # excluding paths.py because I like my paths
                      # long sometimes
)
'''

[tool.bumpversion]
current_version = "0.1.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = false
message = "Bump version: {current_version} → {new_version}"
commit_args = ""
setup_hooks = []
pre_commit_hooks = []
post_commit_hooks = []


[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = false
ignore-setters = false
ignore-overloaded-functions = false
fail-under = 80
# example values
exclude = ["setup.py", "docs", "build"]
# example regex
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
ext = []
# possible values: sphinx (default), google
# TODO(shivashankarv): Causes an error for some reason.
#style = sphinx
# possible values: 0 (minimal output), 1 (-v), 2 (-vv)
verbose = 0
quiet = false
whitelist-regex = []
color = true
omit-covered-files = false
# output file logation
generate-badge = "."
badge-format = "svg"

[tool.isort]
profile = "black"

#TODO: Add pytest-cov to github worklfow in a separate issue. This would likely #require `pip install pytest-cov` and this commented block should work as is.

#[tool.pytest.ini_options]
#addopts = "-v --cov --cov-report term-missing"
#testpaths = [
#    "tests",
#]

[tool.setuptools.packages.find]
where = ["src"]
#this might be useful if you have multiple folders in "src" folder
#include = ["pkg*"]  # alternatively: `exclude = ["additional*"]`
#namespaces = false

[project]
name = "dataproduct-builder"
version = "0.1.0"
dependencies = [
    'pandas>=1.5.0',
    'numpy>=1.21.0',
    'typer>=0.9.0',
    'pyyaml>=6.0',
    'openpyxl>=3.0.0',
    'scikit-learn>=1.0.0',
    'scipy>=1.8.0',
    'matplotlib>=3.5.0',
    'seaborn>=0.11.0',
    'tqdm>=4.64.0',
    'joblib>=1.1.0',
    'jupyterlab>=3.0.0',
    'pytest>=7.0.0',
    'build>=0.8.0',
    'twine>=4.0.0',
]
authors = [
    { name="FogPharma Data Science Team", email="nmathai@parabilismed.com" },
]
description = "A comprehensive data processing pipeline for peptide sequence analysis and general data manipulation"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: cheminformatics",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/FogPharma/dataproduct_builder/"
Issues = "https://github.com/FogPharma/dataproduct_builder/issues"
Repository = "https://github.com/FogPharma/dataproduct_builder.git"

[project.scripts]
dataproduct-builder = "dataproduct_builder.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
]
